Remove cats instances
